[RDY] AnimView compliance fixes

+++ AnimView/th.h
@@ -35,6 +35,7 @@ SOFTWARE.
 */
 
 #pragma once
+#include "config.h"
 #include <wx/string.h>
 #include <wx/file.h>
 #include <wx/image.h>
@@ -117,7 +118,7 @@ public:
         if(0 <= iLayer && iLayer < 13)
             for(int iId = 0; iId < 32; ++iId)
             {
-                if((m_iMask[iLayer] & (1 << iId)) != 0)
+                if((m_iMask[iLayer] & (static_cast<std::uint32_t>(1) << iId)) != 0)
                     return true;
             }
         return false;
